# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shared Accounts Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Accounts
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List or search customers by partial name or phone
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "409":
          description: Phone already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer detail, subscriptions, and total spent
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "409":
          description: Phone already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer without subscriptions
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Customer deleted
        "409":
          description: Customer has subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Sales
...
  /platform/orgs/{id}/suspend:
    post:
      operationId: suspendPlatformOrg
      tags: [platform]
      summary: Suspend an organization
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Suspended organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformOrg"
  /platform/orgs/{id}/unsuspend:
    post:
      operationId: unsuspendPlatformOrg
      tags: [platform]
      summary: Unsuspend an organization
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Unsuspended organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformOrg"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Product:
      type: object
      required:
        [
          id,
          name,
          service,
          defaultCapacity,
          defaultDurationDays,
          defaultPrice,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        service:
          type: string
        defaultCapacity:
          type: integer
        defaultDurationDays:
          type: integer
        defaultPrice:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required:
        [name, service, defaultCapacity, defaultDurationDays, defaultPrice]
      properties:
        name:
          type: string
          minLength: 1
        service:
          type: string
          minLength: 1
        defaultCapacity:
          type: integer
          minimum: 1
        defaultDurationDays:
          type: integer
          minimum: 1
        defaultPrice:
          type: number
          minimum: 0
        notes:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        service:
          type: string
          minLength: 1
        defaultCapacity:
          type: integer
          minimum: 1
        defaultDurationDays:
          type: integer
          minimum: 1
        defaultPrice:
          type: number
          minimum: 0
        notes:
          type: ["string", "null"]
...
    Customer:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
          format: email
        notes:
          type: ["string", "null"]
    CustomerUpdate:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
          format: email
        notes:
          type: ["string", "null"]
    CustomerDetail:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          required: [totalSpent, subscriptions]
          properties:
            totalSpent:
              type: number
            subscriptions:
              type: array
              items:
                $ref: "#/components/schemas/SubscriptionSummary"
    SubscriptionStatus:
      type: string
      enum: [active, expired, cancelled]
...
    SubscriptionSummary:
      type: object
      required:
        [
          id,
          customerId,
          customerName,
          slotId,
          slotIndex,
          accountId,
          accountLabel,
          productId,
          productName,
          startDate,
          expiryDate,
          price,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        slotId:
          type: integer
        slotIndex:
          type: integer
        accountId:
          type: integer
        accountLabel:
          type: string
        productId:
          type: integer
        productName:
          type: string
        startDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        price:
          type: number
        status:
          $ref: "#/components/schemas/SubscriptionStatus"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    PlatformOrg:
      type: object
      required: [id, name, status, createdAt, usersCount, productsCount, accountsCount, customersCount, subscriptionsCount, paymentsCount]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
        ownerEmail:
          type: ["string", "null"]
          format: email
        usersCount:
          type: integer
        productsCount:
          type: integer
        accountsCount:
          type: integer
        customersCount:
          type: integer
        subscriptionsCount:
          type: integer
        paymentsCount:
          type: integer